Sat, Oct 21, 2023, 10:20 PM +07:00

+++ cs50p/week3/fuel/fuel.py
@@ -2,7 +2,7 @@
 while True:
     x, y = input("Fraction: ").split("/")
     try:
-        z = float(x) / float(y)
+        z = int(x) / int(y)
     except Exception:
         pass
     else:
